refactor : Rename method & variables in InputValidator(old ExceptionHandler)
